refactor(frontend): add missing index.ts
