+++ src/core/auth/index.ts
@@ -1,1 +1,2 @@
 export { AuthGuard } from './AuthGuard';
+export { useAuthGuardUser } from './UserContext';
