+++ src/pages/criteria-page/index.ts
@@ -1,0 +1,1 @@
+export { CriteriaPage } from './CriteriaPage';
