+++ src/shared/layouts/dashboard-layouts/DashboardLayout.stories.tsx
@@ -1,9 +1,8 @@
 import { i18n } from '@lingui/core';
 import { storiesOf } from '@storybook/react';
 import React from 'react';
-import { themeDecorator } from 'src/stories/decorators';
-import { routerDecorator } from 'src/stories/decorators/routerDecorator';
-import { Lorem } from 'src/stories/helpers/Lorem';
+import { routerDecorator, themeDecorator } from 'src/stories/decorators';
+import { Lorem } from 'src/stories/helpers';
 import { Brand, BrandRegion, ContentRegion, DashboardLayout, NavBarMenu, NavbarRegion } from '.';
 import { MenuItem } from './menu/types';
