+++ src/shared/sticky-action-bar/index.ts
@@ -1,0 +1,1 @@
+export { StickyActionBar } from './StickyActionBar';
