+++ src/stories/decorators/index.ts
@@ -1,2 +1,3 @@
 export { relayDecorator } from './RelayDecorator';
 export { themeDecorator } from './themeDecorator';
+export { routerDecorator } from './routerDecorator';
