+++ src/stories/direction-state.ts
@@ -1,3 +1,3 @@
-import { DirectionState } from './helpers/DirectionState';
+import { DirectionState } from './helpers';
 
 export const directionState = new DirectionState();
